+++ CHANGES.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2026-08-20
 * Sped up the whole pipeline a lot: the `notebooks/methanol_and_iron.ipynb` case
   study goes from 613 s to 44 s end to end
   ([#64](https://github.com/RWTH-LTT/optimex/pull/64))